Refactor: Rename and organize helpers and types
